[step3] fix, feature: 코드 일부 번경 ( 여러 변경이 있어 아래 기술..)
1. ResultView 생성
2. 패키지 분리 Car 관련, view 관련
3. Driver getMoveResult 응답 변경
4. 테스트 작성
