+++ README.md
@@ -38,3 +38,4 @@
 4. 자동차 전진 혹은 멈춤
     - [x] 자동차 전진 혹은 멈춤의 결과를 구분한다.
 5. 화면 출력
+    - [x] 응답뷰 : 자동차 이동 결과를 출력한다
