+++ src/main/kotlin/step3/racingcar/view/ResultView.kt
@@ -1,0 +1,12 @@
+package step3.racingcar.view
+
+object ResultView {
+    private const val DASH_VIEW = "-"
+
+    fun show(result: Int) {
+        repeat(result) {
+            print(DASH_VIEW)
+        }
+        println()
+    }
+}
