+++ src/test/kotlin/step3/racingcar/view/ViewTest.kt
@@ -1,0 +1,12 @@
+package step3.racingcar.view
+
+import org.junit.jupiter.api.Test
+
+class ViewTest {
+
+    @Test
+    fun `ResultView 객체를 생성, 입력값 만큼 -를 print한다`() {
+        val repeat = 5
+        ResultView.show(5)
+    }
+}
